--- Home.jsx
@@ -1,14 +1,0 @@
-import Code from './src/Code';
-import Result from './Result';
-import Header from './src/Header';
-
-const Home = () => {
-    return (
-        <>
-            <Header />
-            <Code />
-            <Result />
-        </>
-    )
-}
-export default Home;
