+++ src/Code.jsx
@@ -1,0 +1,44 @@
+import { useContext } from 'react';
+//components
+import Editor from './Editor';
+import { Box, styled } from '@mui/material';
+import { DataContext } from './DataProvider';
+
+const Container = styled(Box)`
+    background-color: #060606;
+    height: 50vh;
+    display: flex;
+`
+const Code = () => {
+    const { html, css, js, setHtml, setCss, setJs } = useContext(DataContext);
+    return (
+        <Container>
+            <Editor 
+                language="xml"
+                heading="HTML"
+                value={html}
+                onChange={setHtml}
+                icon='/'
+                color='#FF3C41'
+            />
+            <Editor 
+                language="css"
+                heading="CSS"
+                value={css}
+                onChange={setCss}
+                icon='*'
+                color='#0EBEFF'
+            />
+            <Editor 
+                language="javascript"
+                heading="JS"
+                value={js}
+                onChange={setJs}
+                icon='( )'
+                color='#FCD000'
+            />
+        </Container>
+    )
+}
+
+export default Code;
