+++ src/DataProvider.jsx
@@ -1,0 +1,20 @@
+import { createContext, useState } from 'react';
+export const DataContext = createContext(null);
+const DataProvider = ({ children }) => {
+    const [html, setHtml] = useState('');
+    const [js, setJs] = useState('');
+    const [css, setCss] = useState('');
+    return (
+        <DataContext.Provider value={{
+            html,
+            setHtml,
+            css,
+            setCss,
+            js,
+            setJs
+        }}>
+            {children}
+        </DataContext.Provider>
+    )
+}
+export default DataProvider;
